requirements.txt
beautifulsoup4==4.12.2
numpy==1.24.4
pytesseract==0.3.10
requests==2.31.0
opencv-python==4.8.1.78
gensim==4.3.2
nltk==3.8.1
pywebview==4.3.3
bs4==0.0.1
tensorflow==2.10.1
pystray==0.19.5
pillow==10.0.1